add basic reference provider
