+++ packages/extension/src/parts/ExtensionHost/ExtensionHostDefinitionProviderTypeScript.js
@@ -1,3 +1,4 @@
+import * as GetDefinitionFromTsResult from '../GetDefinitionFromTsResult/GetDefinitionFromTsResult.js'
 import * as Position from '../Position/Position.js'
 import * as Rpc from '../Rpc/Rpc.js'
 import * as TextDocumentSync from '../TextDocumentSync/TextDocumentSync.js'
@@ -10,20 +11,11 @@ export const languageId = 'typescript'
 export const provideDefinition = async (textDocument, offset) => {
   const uri = textDocument.uri
   const tsPosition = Position.getTsPosition(textDocument, offset)
-  console.log({ tsPosition })
   await TextDocumentSync.openTextDocuments([textDocument])
-  const definition = await Rpc.invoke(
-    'Definition.getDefinition',
-    uri,
-    tsPosition
+  const tsResult = await Rpc.invoke('Definition.getDefinition', uri, tsPosition)
+  const definition = GetDefinitionFromTsResult.getDefinitionFromTsResult(
+    textDocument,
+    tsResult
   )
-  console.log({ definition })
-  // TODO
-  // const tsResult = await TsServerRequests.definition({
-  //   file: textDocument.uri,
-  //   line: tsPosition.line,
-  //   offset: tsPosition.offset,
-  // })
-  // const definition = getDefinitionFromTsResult(textDocument, tsResult)
-  return undefined
+  return definition
 }
