+++ packages/extension/src/parts/ExtensionHost/ExtensionHostReferenceProviderTypeScript.js
@@ -1,29 +1,9 @@
-import * as TsServerRequests from '../TsServerRequests/TsServerRequests.js'
+import * as GetReferencesFromTsResult from '../GetReferencesFromTsResult/GetReferencesFromTsResult.js'
 import * as Position from '../Position/Position.js'
+import * as Rpc from '../Rpc/Rpc.js'
+import * as TextDocumentSync from '../TextDocumentSync/TextDocumentSync.js'
 
 export const languageId = 'typescript'
-
-/**
- *
- * @param {vscode.TextDocument} textDocument
- * @param {import('typescript/lib/protocol').ReferencesResponseBody} tsResult
- * @returns {readonly vscode.Reference[]}
- */
-const getReferencesFromTsResult = (textDocument, tsResult) => {
-  const references = []
-  for (const ref of tsResult.refs) {
-    const startOffset = Position.getOffset(textDocument, ref.start)
-    const endOffset = Position.getOffset(textDocument, ref.end)
-    references.push({
-      uri: ref.file,
-      startOffset,
-      endOffset,
-      lineText: ref.lineText,
-    })
-  }
-  console.log(tsResult.refs)
-  return references
-}
 
 // TODO should this function return positions or offsets?
 // when it returns offset, need to convert it to position anyway for references view
@@ -33,13 +13,16 @@ const getReferencesFromTsResult = (textDocument, tsResult) => {
  * @type{vscode.ReferenceProvider['provideReferences']}
  */
 export const provideReferences = async (textDocument, offset) => {
+  await TextDocumentSync.openTextDocuments([textDocument])
   const tsPosition = Position.getTsPosition(textDocument, offset)
-  const tsResult = await TsServerRequests.references({
+  const tsResult = await Rpc.invoke('References.getReferences', {
     file: textDocument.uri,
     line: tsPosition.line,
     offset: tsPosition.offset,
   })
-  console.log({ tsResult })
-  const references = getReferencesFromTsResult(textDocument, tsResult)
+  const references = GetReferencesFromTsResult.getReferencesFromTsResult(
+    textDocument,
+    tsResult
+  )
   return references
 }
