+++ packages/extension/src/parts/Providers/Providers.js
@@ -1,2 +1,3 @@
 export * as CompletionProvider from '../ExtensionHost/ExtensionHostCompletionProviderTypeScript.js'
 export * as DefinitionProvider from '../ExtensionHost/ExtensionHostDefinitionProviderTypeScript.js'
+export * as ReferenceProvider from '../ExtensionHost/ExtensionHostReferenceProviderTypeScript.js'
