+++ packages/extension/src/parts/RegisterProviders/RegisterProviders.js
@@ -7,6 +7,10 @@ const registerProvider = (provider) => {
     vscode.registerDefinitionProvider(provider)
     return
   }
+  if ('provideReferences' in provider) {
+    vscode.registerReferenceProvider(provider)
+    return
+  }
 }
 
 export const registerProviders = (providers) => {
