+++ packages/node/src/parts/Refererences/References.js
@@ -1,0 +1,6 @@
+import * as TsServerRequests from '../TsServerRequests/TsServerRequests.js'
+
+export const getReferences = async (params) => {
+  const tsResult = await TsServerRequests.definition(params)
+  return tsResult
+}
